requirements.txt
# ── Core API ────────────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9      
pydantic>=2.0.0             

# ── Frontend ─────────────────────────────────────────────────────────────────
streamlit>=1.33.0
requests>=2.31.0

# ── Embeddings & Vector Store ────────────────────────────────────────────────
# Pure-numpy TF-IDF — no FAISS or external vector DB required
numpy>=1.26.0

# ── LLM Integration (Groq) ──────────────────────────────────────────────────
groq>=1.1.0
python-dotenv>=1.0.0

# ── Embeddings (pure-numpy TF-IDF, no extra packages needed) ────────────────
# scikit-learn is optional; install if disk space allows: pip install scikit-learn

# ── Dev / Testing ────────────────────────────────────────────────────────────
pytest>=8.2.0
httpx>=0.27.0                
